job_matcher_system/requirements.txt
# Core dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# NLP & ML
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
nltk>=3.8.0

# Data processing
pandas>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Logging & Utils
structlog>=23.0.0
tqdm>=4.66.0
joblib>=1.3.0

# Optional: For advanced scraping
fake-useragent>=1.4.0
